backend/requirements.txt
django
djangorestframework
django-cors-headers
supabase
python-dotenv
psycopg2-binary
dj-database-url
